+++ src/mplchart/expressions/prices.py
@@ -2,13 +2,19 @@
 
 import polars as pl
 
-from .prelude import wrap_expression, HIGH, LOW, CLOSE
+from .prelude import wrap_expression, OPEN, HIGH, LOW, CLOSE
 
 
 @wrap_expression
-def MIDPRICE(*, high: pl.Expr = HIGH, low: pl.Expr = LOW) -> pl.Expr:
+def MEDPRICE(*, high: pl.Expr = HIGH, low: pl.Expr = LOW) -> pl.Expr:
     """Median Price (HL/2)"""
     return (high + low) / 2
+
+
+@wrap_expression
+def AVGPRICE(*, open: pl.Expr = OPEN, high: pl.Expr = HIGH, low: pl.Expr = LOW, close: pl.Expr = CLOSE) -> pl.Expr:
+    """Average Price (OHLC/4)"""
+    return (open + high + low + close) / 4
 
 
 @wrap_expression